requirements.txt
pandas
numpy
Flask
mysql-connector-python

